distinguish between smaller font for InnerAnchor on TopNav & larger font for SideNav otherwise, share all styles
