+++ src/components/Header/SideNav/sideNavItemPadding.js
@@ -1,8 +1,8 @@
 import { css } from 'styled-components'
 import remcalc from 'remcalc'
 
-const itemPadding = css`
+const sideNavItemPadding = css`
   padding: ${remcalc('13')} ${remcalc('38')} ${remcalc('11')} ${remcalc('24')};
 `
 
-export default itemPadding
+export default sideNavItemPadding
